Fixed Problems with Avg Gpa Change, only outlier removal remains
There were some issues with the avg GPA change calculations where the method I was using to get the index of the quarter after the internship was not working, and I forgot the index was for After the internship and not Before the internship.

Now the gpa differences and averages are coming out correctly, but some outliers are skewing the gpa diff average.